fix: input no longer case sensitive
an extra action has been added when converting user input into variables
to account for the possibility of a different case sensitivity. The way
I had it set up originally, when a user was asked if they want to
covnvert a number or not, they would need to enter a lowercase 'y' or
'n'. Similarly, when inputting hex numbers 'A-F', the letters needed to
be capital, otherwise the program would not recognize the number as
valid. An operation has been added in both cases to ensure the case
sensitivity does not matter. This eliminates an annoying little quirk
that may confuse a user.